requirements.txt
aerich==0.7.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-property==0.2.2
asyncclick==8.1.7.2
babel==2.17.0
bcrypt==4.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
deprecation==2.1.0
dictdiffer==0.9.0
ecdsa==0.19.0
fastapi==0.115.6
fastapi_keycloak==1.0.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==3.0.2
numpy==2.2.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pypika-tortoise==0.3.2
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-keycloak==4.7.3
python-multipart==0.0.19
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tomli_w==1.2.0
tomlkit==0.13.2
tortoise-orm==0.22.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wheel==0.44.0
